% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_experiments.R
\name{add_experiments}
\alias{add_experiments}
\title{Adding self-processed data to existing ones}
\usage{
add_experiments(nData, procData, ExpInfoTable)
}
\arguments{
\item{nData}{A list of objects returned by \code{multi_norm_and_annot}}

\item{procData}{Matrix or list of matrices}

\item{ExpInfoTable}{}
}
\value{
List of all matrices with normalized expression values
}
\description{
\code{add_experiments} function enables user to add processed data
}
\details{
Blablabla
}
\examples{
\dontrun{
### download, load, normalize and annotate data from ArrayExpress

# download
to_download = c("E-GEOD-21066", "E-MTAB-966")
data = downloadAE(to_download, getwd())

# load
platforms = c("Affymetrix", "Agilent")
loaded_experiments = load_multi_data(data, platforms)

# process
processed_data = multi_norm_and_annot(loaded_experiments[[1]], platforms)
}

}
