% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ranking.R
\name{create_ranking}
\alias{create_ranking}
\title{Create a stability ranking for genes or miRNAs}
\usage{
create_ranking(all_exp_data, all_uniq_samples, miRNA = FALSE)
}
\arguments{
\item{all_exp_data}{List of expression matrices. The first element of the output of \code{rep_elim} function.}

\item{all_uniq_samples}{List of character matrices with information about unique samples in each experiment. The second element
of the output of \code{rep_elim} function.}

\item{miRNA}{Logical value indicating if stability ranking should be performed for miRNA or genes. Default value is FALSE what means
that ranking will be created for genes.}
}
\value{
Function returns a list. It's content depends on miRNA argument.

If miRNA == FALSE, the list contains as folows:
\itemize{
\item $GeneRanking: stability ranking for genes
\item $controls: pairs of treated and control samples
\item $FC_data: matrices with fold changes
\item $samples: names of treated samples
\item $EntrezId: all Entrez Ids
}
else if miRNA == TRUE the list contains as folows:
\itemize{
\item $miRNA_ranking: stability ranking for miRNAs
\item $controls: pairs of treated and control samples
\item $FC_data: matrices with fold changes
\item $samples: names of treated samples
\item $MIMATids: all MIMAT ids
}
}
\description{
\code{create_ranking} function creates stability ranking of genes or miRNAs expression levels.
}
\details{
Note that \code{create_ranking} works properly only if arguments are the output of \code{rep_elim} function so you should use it
even if you know that there are no replications in your data. If there are no replications in the data \code{rep_elim} function
will just prepared objects to use with \code{create_ranking} function.

The ranking is created with the following steps:
\itemize{
\item Z-scores are calculated for each expression value for each gene within an experiment.
\item For each treated sample a control is founded and logarithmic fold change calculated.
\item For each gene within an experiment gene score is calculated as ratio between logarithmic control expression
and logarithmic fold change.
\item Stability index for each gene is calculated as a weighed mean of gene scores from all experiments.
}

Fold change in time for chosen genes or miRNAs could be plotted with \code{lFC_in_time} function.
You can also generate summary for used experimental conditions with \code{get_info_about_used_exp} function.

It should be made into consideration that reliable ranking can be produced only with reasonable number of microarray data.
}
\examples{
\dontrun{
##### Create stability ranking for genes

# download data from ArrayExpress database
to_download = c("E-GEOD-67309", "E-MTAB-966")
my_data = downloadAE(to_download, getwd())

# load data
platforms = c("Affymetrix", "Agilent")
loaded_data = load_multi_data(my_data, platforms)

# normalize and annotate
norm_data = multi_norm_and_annot(loaded_data$raw_expression_data, platforms)

# prepare tables for rep_elim function as shown in details
path_to_tables = system.file("extdata", "tables_ex3.rds", package = "FindReference")
my_tables = readRDS(path_to_tables)

# eliminate replications and prepare object for create_ranking function
no_rep_data = rep_elim(norm_data, my_tables)

# create ranking
gene_ranking = create_ranking(no_rep$noRepData, no_rep$uniqSamples, miRNA = FALSE)

##### Create stability ranking for miRNAs

# download data from ArrayExpress database
datamiRNA = downloadAE("E-MTAB-5197", "/home/emarek/")

# prepare table as shown in details load_miRNA help page
path_to_table = system.file("inst/extdata", "miRNA_ex1.rds", package = "FindReference")
my_table = readRDS(path_to_table)

# load data
loaded_data = load_miRNA(my_table, datamiRNA[[1]]$path)

# normalize and annotate data
norm_data = norm_and_annot_miRNA(loaded_data)

# eliminate replications and prepare object for create_ranking function
no_rep_data = rep_elim(norm_data, my_table)

# create ranking
miRNA_ranking = create_ranking(no_rep$noRepData, no_rep$uniqSamples, miRNA = TRUE)
}

}
\seealso{
\code{\link{rep_elim}}, \code{\link{lFC_in_time}}, \code{\link{get_info_about_used_exp}}
}
