% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAE.R
\name{downloadAE}
\alias{downloadAE}
\title{Downloading microarray data from ArrayExpress database}
\usage{
downloadAE(ExpInfoTable, path)
}
\arguments{
\item{path}{A character indicating directory where data should be downloaded and extracted.}

\item{ExpIds}{A data frame with two columns named Experiment and RawData.
The first column should contain experiment identifiers from ArrayExpress database to download. In the RawData column there should be
a character "T" or "F" for each experiment, where "T" means that downloaded data should be raw type and "F" for processed type.}
}
\value{
Function returns a list of lists where the first one contains names of downloaded files (the output from getAE function from ArrayExpress package)
and saves it as dataAE.rds. In the second list .sdrf files are stored in which there are information about samples and experiment conditions.
The third element of returned list is a data.frame with information about platform for each experiment.
}
\description{
\code{downloadAE} function downloads microarray data from ArrayExpress database basing on experiment identifiers. Then it identifies
microarray platform on which the experiment was performed using downloaded .sdrf file.
}
\details{
The output of this function is needed later in \code{load_multi_data} and \code{load_data} functions and that's why it is also
saved in dataAE.rds which could be easy loaded with \code{readRDS}.
}
\examples{
\dontrun{downloadAE(c("E-GEOD-65292", "E-TABM-90"), getwd())}

\dontrun{experiments_ids = c("E-GEOD-65292", "E-TABM-90")
downloadAE(experiments_ids, getwd())}

}
