% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_about_used_exp.R
\name{get_info_about_used_exp}
\alias{get_info_about_used_exp}
\title{Summarize information about experiments used to create ranking}
\usage{
get_info_about_used_exp(encoded_info, txt = TRUE, csv = TRUE,
  path = getwd())
}
\arguments{
\item{encoded_info}{List obtained with \code{create_ranking} function.}

\item{txt}{Logical indicating if .txt files with summaries should be exported into txt files. Default value is TRUE.}

\item{csv}{Logical indicating if .csv files with summaries should be exported into txt files. Default value is TRUE.}

\item{path}{Character indicating the directory where files with summaries should be exported. Default value is current directory.}
}
\value{
Function doesn't return a value. Summaries of used cell lines, doses and times after treating are exported into .txt, .csv
files or both.
}
\description{
\code{get_info_about_used_exp} function summarizes conditions under which experiments used to create ranking where performed.
It lists and counts used cell lines, doses and times after treatment which could be useful to see for example if some conditions
where overrepresented.
}
\examples{
\dontrun{
##### Create stability ranking for genes

# download data from ArrayExpress database
to_download = c("E-GEOD-67309", "E-MTAB-966")
my_data = downloadAE(to_download, getwd())

# load data
platforms = c("Affymetrix", "Agilent")
loaded_data = load_multi_data(my_data, platforms)

# normalize and annotate
norm_data = multi_norm_and_annot(loaded_data$raw_expression_data, platforms)

# prepare tables for rep_elim function as shown in details
path_to_tables = system.file("inst/extdata", "tables_ex3.rds", package = "FindReference")
my_tables = readRDS(path_to_tables)

# eliminate replications and prepare object for create_ranking function
no_rep_data = rep_elim(norm_data, my_tables)

# create ranking
gene_ranking = create_ranking(no_rep$noRepData, no_rep$uniqSamples, miRNA = FALSE)

# summarise experimental conditions - export just .csv files in current directory
get_info_about_used_exp(gene_ranking, txt=FALSE, csv=TRUE)
}

}
\seealso{
\code{\link{create_ranking}}
}
