% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lFC_in_time.R
\name{lFC_in_time}
\alias{lFC_in_time}
\title{Plot fold change in time}
\usage{
lFC_in_time(genes_to_valid, dane_ranking_z_scores, scale = "linear",
  miRNA = FALSE)
}
\arguments{
\item{genes_to_valid}{Character vector with gene symbols or MIMAT ids (for miRNA) which fold change in time should be plotted.}

\item{dane_ranking_z_scores}{A list - output of \code{create_ranking} function.}

\item{scale}{A character indicating the scale of plot. Could be 'linear' for fold change or 'log' for logarithmic fold change.
Default value is 'linear'.}

\item{miRNA}{Logical indicating if plot is created for gene or miRNA ranking.}
}
\value{
Function returns a ggplot object which could be plotted with \code{plot} function as in the examples.
}
\description{
\code{lFC_in_time} function could be used to plot fold change in time for chosen genes or miRNAs by gene symbol or MIMAT ids respectively.
}
\examples{
\dontrun{
##### Create stability ranking for genes

# download data from ArrayExpress database
to_download = c("E-GEOD-67309", "E-MTAB-966")
my_data = downloadAE(to_download, getwd())

# load data
platforms = c("Affymetrix", "Agilent")
loaded_data = load_multi_data(my_data, platforms)

# normalize and annotate
norm_data = multi_norm_and_annot(loaded_data$raw_expression_data, platforms)

# prepare tables for rep_elim function as shown in details
path_to_tables = system.file("inst/extdata", "tables_ex3.rds", package = "FindReference")
my_tables = readRDS(path_to_tables)

# eliminate replications and prepare object for create_ranking function
no_rep_data = rep_elim(norm_data, my_tables)

# create ranking
gene_ranking = create_ranking(no_rep$noRepData, no_rep$uniqSamples, miRNA = FALSE)

# plot fold change in time for some genes
genes_to_plot = c('GAPDH', 'ACTB', 'LYPLA2', 'B2M', 'TP53')
genes_FC_in_time = lFC_in_time(genes_to_plot, gene_ranking)

##### Create stability ranking for miRNAs

# download data from ArrayExpress database
datamiRNA = downloadAE("E-MTAB-5197", "/home/emarek/")

# prepare table as shown in details load_miRNA help page
path_to_table = system.file("inst/extdata", "miRNA_ex1.rds", package = "FindReference")
my_table = readRDS(path_to_table)

# load data
loaded_data = load_miRNA(my_table, datamiRNA[[1]]$path)

# normalize and annotate data
norm_data = norm_and_annot_miRNA(loaded_data)

# eliminate replications and prepare object for create_ranking function
no_rep_data = rep_elim(norm_data, my_table)

# create ranking
miRNA_ranking = create_ranking(no_rep$noRepData, no_rep$uniqSamples, miRNA = TRUE)

# plot fold change in time for the most stable miRNAs
miRNA_to_plot = miRNA_ranking$miRNA_ranking[1:9, 'ID']
genes_FC_in_time = lFC_in_time(miRNA_to_plot, miRNA_ranking, scale = 'linear', miRNA = TRUE)
}

}
\seealso{
\code{\link{create_ranking}}
}
