% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loading microarray data from single experiment downloaded from ArrayExpress database}
\usage{
load_data(dane, ExpInfoTable, sdrfFile)
}
\arguments{
\item{dane}{A list from downloadAE function.}

\item{platforma}{Experiment's platform, could be 'Affymetrix' or 'Agilent'.}

\item{donotread}{A character vector indicating which microarrays should not be loaded. Default is NA, which means that all
microarrays listed in dane$rawFiles will be loaded.}
}
\value{
Function returns a list with two elements. The first one is an object specific for array platform with raw expression data
(could be marrayRaw or EListRaw for Agilent platform and AffyBatch or ExonFeatureSet for Affymetrix platform).
The second element is a data frame with loaded .sdrf file.
}
\description{
\code{load_data} function loads data for a single microarray experiment downloaded from ArrayExpress database.
}
\details{
This function is designed to load data for a single experiment. For loading many experiments at once you could use \code{load_multi_data}.

Note that only Agilent (one and double colour) and Affymetrix platforms are supported.

Loaded .sdrf file (second element of the output) could be useful for preparing table with experiment information needed for \code{rep_elim}
or \code{multi_rep_elim} functions.
}
\examples{
\dontrun{
### load all microarrays
dane = downloadAE("E-GEOD-21066", getwd())
loaded_experiment = load_data(dane, "Affymetrix")

### exclude some microarrays from loading
dane = downloadAE("E-GEOD-21066", getwd())
micrroarrays_under_conditions_imnot_interested = c("GSM526680.CEL", "GSM526756.CEL")
loaded_experiment = load_data(dane, "Affymetrix",
                         donotread = micrroarrays_under_conditions_imnot_interested)

### when you downloaded data with downloadAE
### but didn't assign the output into variable
downloadAE("E-GEOD-21066", getwd())

# read in data saved by downloadAE function
dane = readRDS('dataAE.rds')
loaded_experiment = load_data(dane, "Affymetrix")

}

}
\seealso{
\code{\link{load_multi_data}}
}
