% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_and_annot_by_platform.R
\name{norm_and_annot_by_platform}
\alias{norm_and_annot_by_platform}
\title{Normalization and annotation of microarray data}
\usage{
norm_and_annot_by_platform(obiekt, platform)
}
\arguments{
\item{obiekt}{An object of class marrayRaw, EListRaw, AffyBatch, ExonFeatureSet or list of these objects with raw expression data.}

\item{platform}{A character indicating experiment's platform, could be 'Affymetrix' or 'Agilent'.}
}
\value{
An expression matrix with normalized and annotated expression values.
}
\description{
\code{norm_and_annot_by_platform} function normalizes and annotates data from a single microarray experiment.
}
\details{
Normalization

For Affymetrix platform rma background correction and normalization is performed. For Agilent platform background correction
is performed by subtracting background values. Used normalization method for this platform is vsn.

Annotation

Annotation for Affybatch object is made using custom cdf file automatically downloaded from brainarray and for ExonFeatureSet
using huex10sttranscriptcluster.db package. Annotation for Agilent platform is made with org.Hs.eg.db package.

The first argument could be a list due to make it possible to process data from an experiment where there are few array types within
one platform (Affymetrix or Agilent). To process data from many experiments at once you should consider using
\code{multi_norm_and_annot} function.
}
\examples{
\dontrun{
### download, load, normalize and annotate data from ArrayExpress
data = downloadAE("E-GEOD-21066", getwd())
loaded_experiment = load_data(data, "Affymetrix")
processed_data = norm_and_annot_by_platform(loaded_experiment[[1]], "Affymetrix")
}

}
\seealso{
\code{\link{multi_norm_and_annot}}
}
