% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_and_annot_miRNA.R
\name{norm_and_annot_miRNA}
\alias{norm_and_annot_miRNA}
\title{Normalization and annotation of microarray miRNA data}
\usage{
norm_and_annot_miRNA(loaded_data)
}
\arguments{
\item{loaded_data}{An object of class uRNAList or a list of objects of this type.}
}
\value{
Function returns an expression matrix (or list of them if argument was a list) with normalized and annotated expression values.
}
\description{
\code{norm_and_annot_miRNA} function loads normalizes and annotates data from miRNA microarray experiments performed on Agilent platform.
}
\details{
Data are rma background corrected and normalized. They are annotated with MIMAT ids.
}
\examples{
\dontrun{
### download data from ArrayExpress database
AEids = c("E-MTAB-5197", "E-GEOD-59862")
datamiRNA = downloadAE(AEids, getwd())

### prepare tables as shown in details, load them and make sure they're character matrices
path_to_tables = system.file("inst/extdata", "miRNA_ex2.rds", package = "FindReference")
my_tables = readRDS(path_to_tables)

### load data

# for data downloaded from ArrayExpress you can easily get the second argument by:
all_paths = sapply(datamiRNA, function(x){x$path})

loaded_data = load_multi_miRNA(my_tables, all_paths)

### normalize and annotate data
norm_data = norm_and_annot_miRNA(loaded_data)
}

}
